UP-4041 Allow guest user to have transient layout
