+++ uportal-war/src/main/java/org/jasig/portal/layout/UserLayoutManagerFactory.java
@@ -21,7 +21,7 @@ package org.jasig.portal.layout;
 
 import org.jasig.portal.IUserProfile;
 import org.jasig.portal.PortalException;
-import org.jasig.portal.layout.immutable.ImmutableUserLayoutManagerWrapper;
+import org.jasig.portal.layout.immutable.ImmutableTransientUserLayoutManagerWrapper;
 import org.jasig.portal.security.IPerson;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.BeanFactory;
@@ -47,24 +47,17 @@ public class UserLayoutManagerFactory implements BeanFactoryAware {
     }
     
     /**
-     * Obtain a regular user layout manager implementation
+     * Obtain a regular user layout manager implementation (which allows transient layout alterations).
+     * The specific layout type depends on whether the user is a guest user.
      *
      * @return an <code>IUserLayoutManager</code> value
      */
     public IUserLayoutManager getUserLayoutManager(IPerson person, IUserProfile profile) throws PortalException {
         final IUserLayoutManager userLayoutManager = (IUserLayoutManager)this.beanFactory.getBean(USER_LAYOUT_MANAGER_PROTOTYPE_BEAN_NAME, person, profile);
-        
-        return new TransientUserLayoutManagerWrapper(userLayoutManager);
-    }
 
-    /**
-     * Returns an immutable version of a user layout manager.
-     *
-     * @param man an <code>IUserLayoutManager</code> value
-     * @return an immutable <code>IUserLayoutManager</code> value
-     * @exception PortalException if an error occurs
-     */
-    public IUserLayoutManager immutableUserLayoutManager(IUserLayoutManager man) throws PortalException {
-        return new ImmutableUserLayoutManagerWrapper(man);
+        if (person.isGuest()) {
+            return new ImmutableTransientUserLayoutManagerWrapper(userLayoutManager);
+        }
+        return new TransientUserLayoutManagerWrapper(userLayoutManager);
     }
 }
